Start redesign for page bg.
