+++ assets/src/edit-story/components/inspector/design/getDesignPanelsForSelection.js
@@ -20,7 +20,6 @@
 import { elementTypes } from '../../../elements';
 import {
   AnimationPanel,
-  BackgroundSizePositionPanel,
   BackgroundOverlayPanel,
   BorderRadiusPanel,
   BorderStylePanel,
@@ -29,7 +28,7 @@ import {
   LinkPanel,
   LayerStylePanel,
   PageAttachmentPanel,
-  PageStylePanel,
+  PageBackgroundPanel,
   ShapeStylePanel,
   SizePositionPanel,
   TextBoxPanel,
@@ -66,8 +65,8 @@ function getDesignPanelsForSelection(elements) {
     const isBackgroundMedia = !elements[0].isDefaultBackground;
     if (isBackgroundMedia) {
       panels.push({
-        type: PanelTypes.BACKGROUND_SIZE_POSITION,
-        Panel: BackgroundSizePositionPanel,
+        type: PanelTypes.PAGE_BACKGROUND,
+        Panel: PageBackgroundPanel,
       });
       panels.push({
         type: PanelTypes.BACKGROUND_OVERLAY,
@@ -94,7 +93,10 @@ function getDesignPanelsForSelection(elements) {
       });
       // In case of default background without media:
     } else {
-      panels.unshift({ type: PanelTypes.PAGE_STYLE, Panel: PageStylePanel });
+      panels.unshift({
+        type: PanelTypes.PAGE_BACKGROUND,
+        Panel: PageBackgroundPanel,
+      });
       // Always display Presets as the first panel for background.
       panels.unshift({
         type: PanelTypes.STYLE_PRESETS,
@@ -114,7 +116,7 @@ function getDesignPanelsForSelection(elements) {
       switch (type) {
         case PanelTypes.ANIMATION:
           return { type, Panel: AnimationPanel };
-        case PanelTypes.BACKGROUND_SIZE_POSITION:
+        case PanelTypes.PAGE_BACKGROUND:
           // Only display when isBackground.
           return null;
         case PanelTypes.COLOR_PRESETS:
