--- assets/src/edit-story/components/panels/design/pageStyle/pageStyle.js
@@ -1,67 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-import { useCallback } from 'react';
-import { __ } from '@web-stories-wp/i18n';
-
-/**
- * Internal dependencies
- */
-import { Color, Row } from '../../../form';
-import { useStory } from '../../../../app';
-import { SimplePanel } from '../../panel';
-import { getColorPickerActions } from '../../shared';
-
-function PageStylePanel() {
-  const {
-    state: { currentPage },
-    actions: { updateCurrentPageProperties },
-  } = useStory();
-  const { backgroundColor } = currentPage;
-
-  const updateBackgroundColor = useCallback(
-    (value) => {
-      updateCurrentPageProperties({ properties: { backgroundColor: value } });
-    },
-    [updateCurrentPageProperties]
-  );
-
-  return (
-    <SimplePanel
-      name="pageStyle"
-      title={__('Page', 'web-stories')}
-      isPersistable={false}
-    >
-      <Row>
-        <Color
-          hasGradient
-          value={backgroundColor}
-          onChange={updateBackgroundColor}
-          label={__('Background color', 'web-stories')}
-          colorPickerActions={getColorPickerActions}
-          hasOpacity={false}
-        />
-      </Row>
-    </SimplePanel>
-  );
-}
-
-PageStylePanel.propTypes = {};
-
-export default PageStylePanel;
