+++ assets/src/edit-story/components/panels/panel/shared/content.js
@@ -31,7 +31,7 @@ import { useContext } from 'use-context-selector';
 import panelContext from '../context';
 
 const Container = styled.div`
-  padding: ${({ padding }) => padding || '10px 16px'};
+  padding: ${({ padding }) => padding || '4px 16px 8px'};
   overflow: auto;
   background-color: ${({ isSecondary, theme }) =>
     isSecondary ? theme.colors.interactiveBg.secondaryNormal : 'transparent'};
